Update MyDate.java
The default constructor, if used as it stands, could result in a date being created with 0 as the day, month, or year. To make sure that the day, month, and year are initialized so that a valid date exists when this constructor is called, we can use initialization block, or chaining constructors, or even by providing default initialization values for the fields themselves.